+++ Gem/Code/Source/RHI/DualSourceBlendingComponent.cpp
@@ -210,9 +210,9 @@ namespace AtomSampleViewer
             drawItem.m_arguments = drawIndexed;
             drawItem.m_pipelineState = m_pipelineState.get();
             drawItem.m_indexBufferView = &m_indexBufferView;
-            drawItem.m_streamBufferViewCount = static_cast<uint32_t>(m_streamBufferViews.size());
+            drawItem.m_streamBufferViewCount = static_cast<uint8_t>(m_streamBufferViews.size());
             drawItem.m_streamBufferViews = m_streamBufferViews.data();
-            drawItem.m_shaderResourceGroupCount = RHI::ArraySize(shaderResourceGroups);
+            drawItem.m_shaderResourceGroupCount = static_cast<uint8_t>(RHI::ArraySize(shaderResourceGroups));
             drawItem.m_shaderResourceGroups = shaderResourceGroups;
 
             // Submit the triangle draw item.
